♻️ change waitingSeconds argument from int to long

+++ src/test/java/com/raksit/example/NotificationReceiver.java
@@ -22,7 +22,7 @@ public class NotificationReceiver {
     this.topic = topic;
   }
 
-  public List<String> poll(int seconds) throws IOException {
+  public List<String> poll(long seconds) throws IOException {
 
     List<String> records = new ArrayList<>();
